【更新】matsunaka DeleteをSenshuRegistに変更
