+++ WebContent/WEB-INF/struts-config.xml
@@ -437,7 +437,7 @@
 			parameter="operation"
 			scope="request">
 			<forward name="search" path="/SeisekiListSearch.do"/>
-			<forward name="delete" path="/SeisekiListDelete.do"/>
+			<forward name="delete" path="/SeisekiListSenshuRegist.do"/>
 		</action>
 		<!-- 検索 -->
 		<action
@@ -449,9 +449,9 @@
 			scope="request">
 			<forward name="success" path="/jsp/seiseki/seisekiList.jsp"/>
 		</action>
-		<!-- 削除(登録) -->
+		<!-- 選手登録 -->
 		<action
-			path="/SeisekiListDelete"
+			path="/SeisekiListSenshuRegist"
 			type="org.springframework.web.struts.DelegatingActionProxy"
 			name="seisekiListForm"
 			validate="true"
